Set up dialogue table
It's still rough and is currently copied from BSFE with the numbers adjusted and a few missing things at the end added. Next up is of course fixing what needs to be fixed, like dealing with control code mishaps. Not touching kanji, so get the kanji switches to emit blanks.